teachify-backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-multipart
pydantic
python-jose[cryptography]
loguru
Pillow
sentence-transformers
faiss-cpu
numpy
boto3
pytest
google-genai
Pillow
passlib[argon2]
azure-identity
pydantic-settings
python-dotenv
PyPDF2
python-docx
pydantic[email]
moviepy
imageio-ffmpeg
